+++ ProcessLayer/Entities/Lookups/LateDeduction.cs
@@ -4,8 +4,8 @@ namespace ProcessLayer.Entities
 {
     public class LateDeduction
     {
-        public short ID { get; set; }
+        public int ID { get; set; }
         public TimeSpan? TimeIn { get; set; }
-        public byte? DeductedHours { get; set; }
+        public int? DeductedHours { get; set; }
     }
 }
