+++ ProcessLayer/Entities/Lookups/LeaveDefaultCredits.cs
@@ -3,7 +3,7 @@
     public class LeaveDefaultCredits
     {
         public int ID { get; set; }
-        public byte LeaveTypeID { get; set; }
+        public int LeaveTypeID { get; set; }
         public float MinYearsInService { get; set; }
         public float? MaxYearsInService { get; set; }
         public float Credits { get; set; }
