+++ ProcessLayer/Entities/Lookups/NonWorkingDays.cs
@@ -8,7 +8,7 @@ namespace ProcessLayer.Entities
         public DateTime? Day { get; set; }
         public TimeSpan? StartTime { get; set; }
         public TimeSpan? EndTime { get; set; }
-        public byte? NonWorkingType { get; set; }
+        public int? NonWorkingType { get; set; }
         public bool? IsGlobal { get; set; }
         public Lookup Type { get; set; }
         public bool? Yearly { get; set; }
